Ability to launch trial ensembles from saved state. Ability to play / replay individual trials parameterized by stored command-sets.
Main motivation for this latter capability is to analyze the simulation's sensitivity
to time fineness. If low, then this can speed up searching the combinatorial space
considerably. Initial tests are promising. For a single orbit simulation of the Tetra
craft, the differences in final bob positions, when going from 67 microsecond
resolution to 10 microsecond, were on the order of centimeters. So if most of our
maneuvers are stable, then we might be able to discover them at low resolution and
refine their description later at high resolution.